requirements.txt
telethon>=1.36,<2.0
requests>=2.31,<3.0
google-genai
python-dotenv>=1.0,<2.0

